changing data type of delay function from unsigned int to unsigned long, else the maximum delay that will work reliably is 4294 seconds.

+++ inc/modules/pelion_time.h
@@ -6,7 +6,7 @@
  */
 
 void
-pelion_delay_us(unsigned int us);
+pelion_delay_us(unsigned long us);
 
 unsigned long
 pelion_get_current_timestamp_us();
